pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inspector-safe"
version = "0.1.0"
description = "A safe, non-destructive auditing tool for authorized security checks"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "yuki xyz"}]
dependencies = [
  "click>=8.0",
  "httpx>=0.23.0",
  "dnspython>=2.3.0",
  "pytest>=7.0"
]
